// orbito-frontend/src/core/api/generated/models/clientDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export interface ClientDto {
  id?: string;
  tenantId?: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  directEmail?: string | null;
  /** @nullable */
  directFirstName?: string | null;
  /** @nullable */
  directLastName?: string | null;
  isActive?: boolean;
  createdAt?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userFirstName?: string | null;
  /** @nullable */
  userLastName?: string | null;
}